+++ pagination/common/src/commonTest/kotlin/dev/inmo/micro_utils/pagination/utils/PaginationReversing.kt
@@ -7,18 +7,19 @@ import kotlin.test.assertEquals
 class PaginationReversingTests {
     @Test
     fun testThatCommonCaseWorksOk() {
-        val pageSize = 2
-        val collectionSize = 8
+        val pageSize = 3
+        val collectionSize = 9
         val pages = calculatePage(collectionSize, pageSize)
 
-        doWithPagination(FirstPagePagination(pageSize)) {
-            val reversed = it.reverse(collectionSize.toLong())
-            assertEquals(Pagination(calculatePage(collectionSize - it.firstIndex - it.size, it.size), it.size), reversed)
-            if (it.page < pages) {
-                it.nextPage()
-            } else {
-                null
-            }
-        }
+        assertEquals(Pagination(-1, pageSize).reverse(collectionSize), Pagination(0, 0))
+
+        val middleFirstIndex = collectionSize / 2 - pageSize / 2
+        val middleLastIndex = middleFirstIndex + pageSize - 1
+        assertEquals(
+            PaginationByIndexes(middleFirstIndex, middleLastIndex).reverse(collectionSize),
+            PaginationByIndexes(middleFirstIndex, middleLastIndex)
+        )
+
+        assertEquals(Pagination(calculatePagesNumber(collectionSize, pageSize), pageSize).reverse(collectionSize), Pagination(0, 0))
     }
 }